# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Firefly Airline",
        "title": "Firefly Airline API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/internal/airports": {
            "get": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Returns all airports, optionally filtered by city.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "List airports",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by city ID",
                        "name": "city_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-array_internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Creates a new airport.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Create an airport",
                "parameters": [
                    {
                        "description": "Airport information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airport.CreateAirportRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/internal/airports/{id}": {
            "get": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Gets an airport by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Get an airport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Airport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AirlineAPIKey": []
                    }
                ],
                "description": "Updates an airport.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airports"
                ],
                "summary": "Update an airport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Airport ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Airport information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domains_airport.UpdateAirportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/firefly-airline_internal_api.APIResponse-internal_domains_airport_AirportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "firefly-airline_internal_api.APIResponse-array_internal_domains_airport_AirportResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_domains_airport.AirportResponse"
                    }
                }
            }
        },
...
        "firefly-airline_internal_api.APIResponse-internal_domains_airport_AirportResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/internal_domains_airport.AirportResponse"
                }
            }
        },
...
        "internal_domains_airport.AirportResponse": {
            "type": "object",
            "properties": {
                "cityId": {
                    "type": "integer",
                    "example": 1
                },
                "code": {
                    "type": "string",
                    "example": "IKA"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Imam Khomeini International Airport"
                }
            }
        },
        "internal_domains_airport.CreateAirportRequest": {
            "type": "object",
            "properties": {
                "cityId": {
                    "type": "integer",
                    "example": 1
                },
                "code": {
                    "type": "string",
                    "example": "IKA"
                },
                "name": {
                    "type": "string",
                    "example": "Imam Khomeini International Airport"
                }
            }
        },
        "internal_domains_airport.UpdateAirportRequest": {
            "type": "object",
            "properties": {
                "cityId": {
                    "type": "integer",
                    "example": 1
                },
                "code": {
                    "type": "string",
                    "example": "IKA"
                },
                "name": {
                    "type": "string",
                    "example": "Imam Khomeini International Airport"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AgencyAPIKey": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "AirlineAPIKey": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}